+++ decidim-proposals/app/commands/decidim/proposals/admin/create_proposal.rb
@@ -49,8 +49,8 @@ module Decidim
 
         def attributes
           {
-            title: form.title,
-            body: form.body,
+            title: form.formatted_title,
+            body: form.formatted_body,
             category: form.category,
             scope: form.scope,
             component: form.component,