+++ decidim-proposals/app/forms/decidim/proposals/proposal_form.rb
@@ -6,6 +6,8 @@ module Decidim
     class ProposalForm < Decidim::Form
       mimic :proposal
 
+      include Decidim::Proposals::Concerns::FormattedAttributes
+
       attribute :title, String
       attribute :body, String
       attribute :address, String